requirements.txt
# HTTP requests for API calls
requests>=2.28.0

# YAML configuration file support
PyYAML>=6.0

# Environment variable loading from .env files
python-dotenv>=0.19.0